Check values

+++ Ambroise_JEE/src/main/java/fr/alten/ambroiseJEE/model/dao/SkillRepository.java
@@ -32,5 +32,5 @@ public interface SkillRepository extends MongoRepository<Skill, Long> {
 	 * @return An Optional with the corresponding skill or not.
 	 * @author Thomas Decamp
 	 */
-	public Optional<Skill> findByNameAndGrade(String name, float grade);
+	public Optional<Skill> findByNameAndGrade(String name, double grade);
 }
